requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
openai==1.109.1
pydantic==2.11.9
python-dotenv==1.1.1
requests==2.32.5
jinja2==3.1.6
